Renaming the 'excpetion' package to 'exception'
